refactor: move step folder up a level
