+++ src/routes/index.js
@@ -2,4 +2,5 @@ export * from "./account";
 export * from "./home";
 export * from "./login";
 export * from "./profile";
-// export * from "./sign-up";
+export * from "./sign-up";
+export * from "./steps";
