+++ src/routes/sign-up/index.js
@@ -1,0 +1,1 @@
+export * from "./sign-up";
