+++ src/routes/sign-up/sign-up.jsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo, useState } from "react";
 import StepWizard from "react-step-wizard";
 import { WizardContext } from "../../contexts/wizard-context";
-import { Steps } from "./steps";
+import { Steps } from "../steps";
 /* ----------------------------- Wizard Context ----------------------------- */
 //context - does this need to go in a separate file
 //does one context work for the entire app
