+++ src/routes/steps/connect-wallet-scan/body.jsx
