+++ src/routes/steps/user-info.jsx
@@ -1,29 +1,28 @@
-import chevron from "../../../../assets/chevron.svg";
-import { Button, Layout } from "../../../../components"; //button needs to be changed when avail
-import { Header } from "../../../../components/layout/header";
-import { Body } from "./body";
+import chevron from "../../assets/chevron.svg";
+import { Button, Layout } from "../../components"; //button needs to be changed when avail
+import { Header } from "../../components/layout/header";
+
 //TODO: Update buttons when available
-//TODO: Hook up buttons and form inputs for choosing peers
-//TODO: Create logic to check array.length of peers to match number of peers
-//TODO: Button disable/enable logic based on if #peers chosen
+//TODO: need to import and add corresponding form when it is created
+//TODO: needs logic to disable/enable button - ?controlled by form validation?
 
 /**
- * Choose Peers Component -
+ * User-Info Component - Sign up Info
  * @author [K. Ehrenclou](https://github.com/kehrenclou)
  * @param onBackClick back click handler prop passed down from parent
  * @param onForwardClick forward click handler prop passed down from parent
+ * @param onConnectWalletClick connect wallet click handler prop passed down from parent
  * @param header expects header components
  * @param body expects body components
  * @param footer expects footer components
  */
 
-export function ChoosePeers({ onBackClick, onForwardClick }) {
+export function UserInfo({ onBackClick, onForwardClick }) {
   return (
-    <Layout //Layout and all props passed here
+    <Layout
       className="mb-32px"
       header={
         <>
-          {/* Layout.Header.Left */}
           <Header.Left>
             <Button
               buttonKind="textOnly"
@@ -34,14 +33,14 @@ export function ChoosePeers({ onBackClick, onForwardClick }) {
           </Header.Left>
           <Header.Center>
             <Header.Title title="Sign up" />
-            <Header.SubTitle subtitle="Step 3/4" />
+            <Header.SubTitle subtitle="Step 1/4" />
           </Header.Center>
         </>
       }
-      body={<Body></Body>}
+      body="insert sign up form component here"
       footer={
         <Button size="lg" buttonKind="primary" onClick={onForwardClick}>
-          Share seed phrase
+          Continue
         </Button>
       } //replace buton with step button
     ></Layout>
